+++ docs/changelog.rst
@@ -4,10 +4,11 @@ Changelog
 Unreleased
 ----------
 
-* User can pass an `Options` instance (from `qiskit-ibm-runtime`)
+* User can pass a `SamplerOptions` instance (from `qiskit-ibm-runtime`)
   via a keyword argument to both an `IBMQBackend` constructor and 
   an instance method `IBMQBackend.process_circuits`.
 * Remove dependency on deprecated qiskit-ibm-provider.
+* Remove support for deprecated "ibmq_qasm_simulator" backend.
 
 0.53.0 (April 2024)
 -------------------